chat/serializers.py
from rest_framework import serializers
from .models import Chat, Messages
from django.contrib.auth.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username')


class ChatSerializer(serializers.ModelSerializer):
    participants = UserSerializer(many=True, read_only=True)

    class Meta:
        model = Chat
        fields = ('id', 'participants')


class MessagesSerializer(serializers.ModelSerializer):
    sender = UserSerializer(read_only=True)

    class Meta:
        model = Messages
        fields = ('id', 'chat', 'sender', 'content', 'timestamp')
