requirements.txt
django
djangorestframework
Pillow
psycopg-binary
